+++ src/commands/install.ts
@@ -1,6 +1,6 @@
 import { execCommand } from '../utils/execCommand';
 
-const packages = [
+const defaultPackages = [
   'expo-icp',
   'canister-manager',
   'expo-crypto-universal',
@@ -15,8 +15,22 @@ const packages = [
   'expo-storage-universal-web',
 ];
 
-export const install = async () => {
-  console.log('Installing necessary packages...');
+const iiIntegrationHelpersPackages = [
+  'expo-icp',
+  'expo-icp-app-connect-helpers',
+  'expo-icp-frontend-helpers',
+  'ii-integration-helpers',
+];
+
+type InstallOptions = {
+  iiIntegrationHelpers?: boolean;
+};
+
+export const install = async (options: InstallOptions) => {
+  console.log('Installing necessary packages...', options);
+  const packages = options.iiIntegrationHelpers
+    ? iiIntegrationHelpersPackages
+    : defaultPackages;
   packages.forEach((pkg) => {
     console.log(`Installing ${pkg}...`);
     execCommand(`npm install ${pkg}`);