+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.11] - 2025-05-16
+
+### Added
+
+- Added --ii-integration-helpers option to install command
+- Added support for installing only II integration helper packages
+- Updated test cases to cover both default and II integration helper installation paths
+
 ## [0.1.10] - 2025-05-16
 
 ### Changed